# labs/backbase/specs/openapi/backbase/oas-v3.0.1/backbase.yaml
openapi: 3.0.1
info:
  title: Backbase Public API
  description: Backbase Public API Specification.
  version: 2.0.0
servers:
  - url: http://localhost:8080/backbase/v1/
externalDocs:
  description: For any query, reach out to us at slack channel '#backbase-support
  url: https://www.backbase.com/
tags:
  - name: Backbase Public API
paths:
  /version:
    get:
      summary: Get API Version
      description: Returns the version of the API.
      operationId: getVersionInfo
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                    example: 1.0.0
  /accounts/{account_number}:
    get:
      tags:
        - Accounts
      summary: Get details of given Account.
      operationId: getAccountInfo
      parameters:
        - name: account_number
          in: path
          required: true
          schema:
            type: integer
            example: 52421344
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInfo'
        '400':
          description: Invalid Account Number
          content:
            application/json:
              example:
                code: BB0100700
                type: External input
                message: Invalid account number provided
                help: Provide valid account number, or reach out to slack channel '#backbase-support'
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: Authorisation error
          content:
            application/json:
              example:
                code: BB0002700
                type: External input
                message: Authorisation failed
                help: Verify you are authorisied to perform the operation
              schema:
                $ref: '#/components/schemas/ApiResponseError'
    put:
      tags:
        - Accounts
      summary: Update an Account
      description: Updates an existing user account.
      operationId: updateAccount
      parameters:
        - name: account_number
          in: path
          description: ID of the account to update
          required: true
          schema:
            type: integer
            example: 52421344
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInfo'
        '400':
          description: Invalid Account Number
          content:
            application/json:
              example:
                code: BB0100700
                type: External input
                message: Invalid account number provided
                help: Provide valid account number, or reach out to slack channel '#backbase-support'
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: Authorisation error
          content:
            application/json:
              example:
                code: BB0002700
                type: External input
                message: Authorisation failed
                help: Verify you are authorisied to perform the operation
              schema:
                $ref: '#/components/schemas/ApiResponseError'
    delete:
      tags:
        - Accounts
      summary: Delete an Account
      description: Deletes an existing user account.
      operationId: deleteAccount
      parameters:
        - name: account_number
          in: path
          description: ID of the account to update
          required: true
          schema:
            type: integer
            example: 52421344
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInfo'
        '400':
          description: Invalid Account Number
          content:
            application/json:
              example:
                code: BB0100700
                type: External input
                message: Invalid account number provided
                help: Provide valid account number, or reach out to slack channel '#backbase-support'
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: Authorisation error
          content:
            application/json:
              example:
                code: BB0002700
                type: External input
                message: Authorisation failed
                help: Verify you are authorisied to perform the operation
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /accounts:
    post:
      tags:
        - Accounts
      summary: Create an Account
      description: Creates a new user account.
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreationInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountInfo'
        '400':
          description: Bad request
          content:
            application/json:
              example:
                code: BB0202701
                type: External input
                message: Account creation criteria not met
                help: Reach out to slack channel '#backbase-support'
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: Authorisation error
          content:
            application/json:
              example:
                code: BB0002700
                type: External input
                message: Authorisation failed
                help: Verify you are authorisied to perform the operation
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '409':
          description: Account already exists
          content:
            application/json:
              example:
                code: BB0202700
                type: External user input
                message: Account already exists
                help: Reach out to slack channel '#backbase-support'
              schema:
                $ref: '#/components/schemas/ApiResponseError'
  /accounts/{account_number}/transactions:
    get:
      tags:
        - AccountTransaction
      summary: Get Transaction Details
      description: Retrieves the transaction details of for a Account.
      operationId: getTransaction
      parameters:
        - name: account_number
          in: path
          description: ID of the account to get transaction
          required: true
          schema:
            type: integer
            example: 52421344
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionInfoCollection'
        '400':
          description: Invalid Account Number
          content:
            application/json:
              example:
                code: BB0100700
                type: External input
                message: Invalid account number provided
                help: Provide valid account number, or reach out to slack channel '#backbase-support'
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: Authorisation error
          content:
            application/json:
              example:
                code: BB0002700
                type: External input
                message: Authorisation failed
                help: Verify you are authorisied to perform the operation
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '404':
          description: No trsanction found
          content:
            application/json:
              example:
                code: BB0300700
                type:
                message: Unable to find resource with given parameters
                help: Reach out to slack channel '#backbase-support'
              schema:
                $ref: '#/components/schemas/ApiResponseError'
    post:
      tags:
        - AccountTransaction
      summary: Create Transaction
      description: Creates a new transaction for a bank account.
      operationId: createTransaction
      parameters:
        - name: account_number
          in: path
          description: ID of the account to create transaction
          required: true
          schema:
            type: integer
            example: 52421344
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionInfo'
        '400':
          description: Bad request
          content:
            application/json:
              example:
                code: BB0202701
                type: External input
                message: Transaction criteria not met
                help: Reach out to slack channel '#backbase-support'
              schema:
                $ref: '#/components/schemas/ApiResponseError'
        '401':
          description: Authorisation error
          content:
            application/json:
              example:
                code: BB0002700
                type: External input
                message: Authorisation failed
                help: Verify you are authorisied to perform the operation
              schema:
                $ref: '#/components/schemas/ApiResponseError'
components:
  schemas:
    AccountInfo:
      type: object
      properties:
        accountNumber:
          type: integer
          example: 52421344
        accountType:
          type: string
          example: Saving
        accountHolderName:
          type: string
          example: BackbaseFinance
        accountHolderPrimaryEmail:
          type: string
          example: backbase@gmail.com
        accountHolderMobile:
          type: string
          example: '+91-9986316924'
        dailyTransactionLimit:
          type: string
          example: 10000
        branchDetails:
          type: object
          items:
            '$ref': '#/components/schemas/BranchInfo'
        additionalProperties:
          type: string
    AccountCreationInput:
      type: object
      properties:
        accountHolderName:
          type: string
          example: BackbaseFinance
        accountHolderPrimaryEmail:
          type: string
          example: backbase@gmail.com
        accountHolderMobile:
          type: string
          example: '+91-9986316924'
        bankName:
          type: string
          example: 'HDFC'
        branchCode:
          type: string
          example: 'HDFCIN000001'
    AccountUpdateInput:
      type: object
      properties:
        dailyTransactionLimit:
          type: integer
          example: 10000
        accountHolderPrimaryEmail:
          type: string
          example: backbase@gmail.com
        accountHolderMobile:
          type: string
          example: '+91-9986316924'
    BranchInfo:
      type: object
      properties:
        bankName:
          type: string
          example: 'HDFC'
        bankSwiftCode:
          type: string
          example: 'HDFC000012'
        bankIfscCode:
          type: string
          example: '121314144'
        bankBranchCode:
          type: string
          example: 'HDFCIN0001'
    TransactionInfo:
      type: object
      properties:
        transactionId:
          type: string
          example: '550e8400-e29b-41d4-a716-446655440000'
        date:
          type: string
          format: date-time
          example: '2017-07-21T17:32:28Z'
        paymentInfo:
          '$ref': '#/components/schemas/PaymentInfo'
    PaymentInfo:
      type: object
      properties:
        amount:
          type: number
          example: 9001.85
        paymentType:
          type: string
          enum: ['CR', 'DB']
        paymentInfo:
          oneOf:
            - $ref: '#/components/schemas/CreditInfo'
            - $ref: '#/components/schemas/DebitInfo'
    TransactionInfoCollection:
      type: array
      items:
        $ref: '#/components/schemas/TransactionInfo'
    TransactionInput:
      type: object
      properties:
        paymentInfo:
          '$ref': '#/components/schemas/PaymentInfo'
        remark:
          type: string
    CreditInfo:
      type: object
      properties:
        senderDetails:
          type: object
          properties:
            bankName:
              type: string
              example: 'IDFC'
            branchCode:
              type: string
              example: 'IDFC002324'
    DebitInfo:
      type: object
      properties:
        receiverDetails:
          type: object
          properties:
            bankName:
              type: string
              example: 'IDFC'
            branchCode:
              type: string
              example: 'IDFC002324'
            purpose:
              type: string
              example: "Investment"
    ApiResponseError:
      type: object
      properties:
        code:
          type: string
          example: BB0600700
        type:
          type: string
          example: External User input
        message:
          type: string
          example: Unable to find resource with given parameters
        help:
          type: string
          example: Reach out to slack channel '#backbase-support'