# labs/backbase/src/main/java/com/backbase/api/common/data/entity/TransactionInfo.java
package com.backbase.api.common.data.entity;

import com.backbase.api.common.data.BackbaseData;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
public class TransactionInfo implements BackbaseData {
    private String transactionId;
    private String date;
    private PaymentInfo paymentInfo;
}